% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{attentionshifting_signal}
\alias{attentionshifting_signal}
\title{Dataset containing th signal for a given electrod (O1) in an experiment using EEG}
\format{A data frame with 120 rows and 819 variables.}
\usage{
data(attentionshifting_signal)
}
\description{
The ERP (averaged EEG signal) for a given electrod (O1) in an experiment in attention shifting sampled at 1024 Hz. The design of the experiment is given in the dataset \link{attentionshifting_design}.
}
\details{
\itemize{
\item ERP (in muV) of the electrod O1 measured from -200 to 600 timeframes before and after the onset of the stimulus.
}
}
