% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_maxT.R
\name{compute_maxT}
\alias{compute_maxT}
\title{The max-T correction}
\usage{
compute_maxT(distribution, alternative)
}
\arguments{
\item{distribution}{A matrix of permuted statistical signal. The first row indicating the observed statistics.}

\item{alternative}{A character string indicating the alternative hypothesis. Default is \code{"greater"}. Choose between \code{"greater"}, \code{"less"} or \code{"two.sided"}.}
}
\description{
Compute the max-T correction given a matrix a permuted statistics.
}
\seealso{
Other multcomp: 
\code{\link{compute_clusterdepth_head}()},
\code{\link{compute_clusterdepth}()},
\code{\link{compute_clustermass}()},
\code{\link{compute_minP}()},
\code{\link{compute_stepdownmaxT}()},
\code{\link{compute_tfce}()},
\code{\link{compute_troendle}()}
}
\concept{multcomp}
