% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_minP.R
\name{compute_minP}
\alias{compute_minP}
\title{The min-P correction}
\usage{
compute_minP(distribution, alternative)
}
\arguments{
\item{distribution}{A matrix of permuted statistical signal. The first row indicating the observed statistics.}

\item{alternative}{A character string indicating the alternative hypothesis. Default is \code{"greater"}. Choose between \code{"greater"}, \code{"less"} or \code{"two.sided"}.}
}
\description{
Compute the min-P correction given a matrix a permuted statistics.
}
\seealso{
Other multcomp: 
\code{\link{compute_clustermass}()},
\code{\link{compute_tfce}()},
\code{\link{compute_troendle}()}
}
\concept{multcomp}
