% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{emergencycost}
\alias{emergencycost}
\title{Dataset of cost of emergency patients.}
\format{A data frame with 176 rows and 5 variables.}
\usage{
data(emergencycost)
}
\description{
Observational data from 176 emergency patients with variables :
}
\details{
\itemize{
\item \code{sex}.
\item \code{age}.
\item \code{insurance} : the type of insurance, private or semi private (\code{semi_private}) or public (\code{public}).
\item \code{LOS} : the length of the stay in days.
\item \code{cost} : the cost in CHF.
}
}
\references{
Heritier, S., Cantoni, E., Copt, S., & Victoria-Feser, M. P. (2009). Robust methods in Biostatistics (Vol. 825). John Wiley & Sons.
}
