% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cluster.R
\name{get_cluster}
\alias{get_cluster}
\title{Get the ID of the cluster on a distribution matrix}
\usage{
get_cluster(distribution, threshold, alternative = "two.sided")
}
\arguments{
\item{distribution}{a matrix of the distribution.}

\item{threshold}{a numeric indicating the threshold.}

\item{alternative}{A character string indicating the alternative hypothesis. Choose between \code{"two.sided"}, \code{"greater"}, \code{"less"}. Default is \code{"two.sided"}.}
}
\value{
a matrix with integer indicating the id of the clusters
}
\description{
Get if of clusters given a threshold
}
\keyword{internal}
