% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_methods.R
\name{plot.clusterlm}
\alias{plot.clusterlm}
\title{Plot cluster or parameters.}
\usage{
\method{plot}{clusterlm}(x, effect = "all", type = "statistic",
  multcomp = "maris_oostenveld", laterality = "bilateral",
  uncorrected = T, ...)
}
\arguments{
\item{x}{A \code{clusterlm} object.}

\item{effect}{A vector of character naming the effects to display. Default is \code{"all"}.}

\item{type}{A character string that specified the values to highlight. \code{"statistic"} or \code{"coef"} are available. Default is \code{"statistic"}}

\item{multcomp}{A character sting specifying the p-value to plot. Default is \code{"maris_oostenveld"}. See \link{clusterlm}.}

\item{laterality}{A character string specifying the laterality of the test when t-test are computed. Avaible options are \code{"right"}, \code{"left"} and \code{"bilateral"}. Default is \code{"bilateral"}.}

\item{uncorrected}{logical. Default is \code{TRUE}. If \code{TRUE}, the uncorrected statistic will be plotted overwise it will plot the corrected statistic. Change for the \code{"ftce"} or the \code{"maris_oostenveld"}, multiple comparisons method.}

\item{...}{further argument pass to plot.}
}
\description{
Plot method for class \code{clusterlm}.
}
