% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.lmperm}
\alias{plot.lmperm}
\title{Plot method for class \code{"lmperm"}.}
\usage{
\method{plot}{lmperm}(x, FUN = density, ...)
}
\arguments{
\item{x}{A \code{"lmperm"} object.}

\item{FUN}{A function to compute the density. Default is \link{density}.}

\item{...}{futher arguments pass to plot.}
}
\description{
Show the density of statistics and the test statistic.
}
\details{
Other argument can be pass to the function : \cr \cr
\code{effect} : a vector of character string indicating the name of the effect to plot. \cr
}
