% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pmat_product.R
\name{Pmat_product}
\alias{Pmat_product}
\title{Multiplies a vector with a Pmat object}
\usage{
Pmat_product(x, P, type = NULL)
}
\arguments{
\item{x}{A a vector or matrix}

\item{P}{A Pmat object}

\item{type}{A character string indicating the type of transformation. This argument need to be specified if P is not of class Pmat.}
}
\value{
A matrix n x np containing the permutated/signflipped vectors.
}
\description{
Multiplies a vector or matrix with a Pmat object
}
