% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_methods.R
\name{print.clusterlm}
\alias{print.clusterlm}
\title{Print \code{clusterlm} object.}
\usage{
\method{print}{clusterlm}(x, laterality = "bilateral", ...)
}
\arguments{
\item{x}{A \code{clusterlm} object.}

\item{laterality}{A character string indicating the laterality of the tests. Choose between \code{"bilateral"}, \code{"right"}, \code{"left"}. Default is \code{"bilateral"}.}

\item{...}{Further arguments pass to \code{print}.}
}
\description{
Display the \code{"maaris_oostenveld"} clusters.
}
