% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_private.R
\name{compute_tfce}
\alias{compute_tfce}
\title{Threshold-Free Cluster-Enhancement correction}
\usage{
compute_tfce(distribution, E, H, ndh)
}
\arguments{
\item{distribution}{A matrix of permuted statistical signal. The first row indicating the observed statistics.}

\item{E}{A scalar that represent the extend parameter of the TFCE transformation. Default is \code{E = 0.5}.}

\item{H}{A scalar that represent the height parameter of the TFCE transformation. Default is \code{H = 1}.}

\item{ndh}{The number of terms in the approximation of the integral.}
}
\description{
Compute the TFCE correction given a matrix a permuted statistical signals.
}
