% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_methods.R
\name{print.clusterlm}
\alias{print.clusterlm}
\title{Print \code{clusterlm} object.}
\usage{
\method{print}{clusterlm}(x, multcomp = NULL, alternative = "two.sided", ...)
}
\arguments{
\item{x}{A \code{clusterlm} object.}

\item{multcomp}{A character string indicating the multiple comparison procedure to print. Default is NULL a print the first multiple comparisons procedure of the \code{clusterlm} object.}

\item{alternative}{A character string indicating the alternative hypothesis. Choose between \code{"two.sided"}, \code{"greater"}, \code{"less"}. Default is \code{"two.sided"}.}

\item{...}{Further arguments pass to \code{print}.}
}
\description{
Display the corrected p-values for each effects. Results of the \code{"clustermass"} procedure.
}
